requirements.txt
oss2
pycryptodome
